Change way of weighting nodes

+++ algorithm/src/main/java/com/agh/bsct/algorithm/services/graph/GraphNode.java
@@ -8,12 +8,12 @@ import java.util.Objects;
 public class GraphNode {
 
     private final long id;
-    private final int weight;
+    private final double weight;
     private GeographicalNodeDTO geographicalNodeDTO;
     private Colour nodeColour;
 
 
-    public GraphNode(Long id, Integer weight, GeographicalNodeDTO geographicalNodeDTO) {
+    public GraphNode(Long id, Double weight, GeographicalNodeDTO geographicalNodeDTO) {
         this.id = id;
         this.weight = weight;
         this.geographicalNodeDTO = geographicalNodeDTO;
@@ -29,7 +29,7 @@ public class GraphNode {
         return id;
     }
 
-    public int getWeight() {
+    public double getWeight() {
         return weight;
     }
